refactor: delete files
