--- test/compile_test.cpp
@@ -1,73 +1,0 @@
-// Copyright (c) 2021 ICHIRO ITS
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-// THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
-#include <gtest/gtest.h>
-#include <rclcpp/rclcpp.hpp>
-#include <shisen_cpp/shisen_cpp.hpp>
-
-#include <memory>
-
-TEST(CompileTest, Interface) {
-  shisen_cpp::CaptureSetting();
-  shisen_cpp::ConfigureCaptureSetting();
-  shisen_cpp::Image();
-}
-
-TEST(CompileTest, Consumer) {
-  try {
-    auto node = std::make_shared<rclcpp::Node>("compile_test");
-
-    {
-      shisen_cpp::CaptureSettingConsumer::Options options;
-      options.camera_prefix = "foo";
-
-      std::make_shared<shisen_cpp::CaptureSettingConsumer>(node, options);
-    }
-
-    {
-      shisen_cpp::ImageConsumer::Options options;
-      options.camera_prefix = "foo";
-
-      std::make_shared<shisen_cpp::ImageConsumer>(node, options);
-    }
-  } catch (...) {
-  }
-}
-
-TEST(CompileTest, Provider) {
-  try {
-    auto node = std::make_shared<rclcpp::Node>("compile_test");
-
-    {
-      shisen_cpp::CaptureSettingProvider::Options options;
-      options.camera_prefix = "foo";
-
-      std::make_shared<shisen_cpp::CaptureSettingProvider>(node, options);
-    }
-
-    {
-      shisen_cpp::ImageProvider::Options options;
-      options.camera_prefix = "foo";
-
-      std::make_shared<shisen_cpp::ImageProvider>(node, options);
-    }
-  } catch (...) {
-  }
-}
